Add files via upload
A number of properties and features of interprocess communication systems are presented, with emphasis on those necessary or desirable in a network environment. The interactlonsbetween these features are examlned, and the consequences of thelr incluslon in a system are explored. Of special interest are the tlme-out feature which forces all system table entries to "die of old age" after they have remained unused for some period of time, and the insertion property which states that it is always possible to design a process which may be Invls~bly inserted into the communication path between any two processes. Though not fled_to any particular system, the discussion concentrates on distributed systems of sequential processes (no interrupts) with no system buffering.